RT In One Weekend - 3. The vec3 Class: add type alias Float = f64

+++ src/main.rs
@@ -1,4 +1,5 @@
 use crate::color::{Color, write_color};
+use crate::vec3::Float;
 
 mod vec3;
 mod color;
@@ -18,8 +19,8 @@ fn main() {
         eprint!("\rScanlines remaining: {} ", j);
         for i in 0..image_width {
             let pixel_color = Color::new(
-                i as f64 / (image_width - 1) as f64,
-                j as f64 / (image_height - 1) as f64,
+                i as Float / (image_width - 1) as Float,
+                j as Float / (image_height - 1) as Float,
                 0.25,
             );
             write_color(&mut std::io::stdout(), &pixel_color).expect("Error");
